+++ Documentation/Changelog.md
@@ -8,6 +8,7 @@ Release date: not released yet
 
 - Added ability to get a `System.Version` based on the given version with the specified number of components set using `System.Version.WithComponents(int)` extension method
 - Added ability to search a sorted `IReadOnlyList<T>` for an element using binary search algorithm implemented by `IReadOnlyList<T>.BinarySearchBy` extension method
+- Added `string.Join<T>(char, IEnumerable<T>)` polyfill method
 
 ### Gapotchenko.FX 2026.3